refactor: change package structure_UIState.kt
