+++ app/src/main/java/campus/tech/kakao/map/viewmodel/MainViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import campus.tech.kakao.map.UIState
 import campus.tech.kakao.map.model.Place
 import campus.tech.kakao.map.repository.PlaceRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
